+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/errorhandlers/StoredRecordChunksErrorHandler.java
@@ -7,6 +7,7 @@ import io.vertx.kafka.client.producer.KafkaHeader;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.dataimport.util.OkapiConnectionParams;
+import org.folio.dataimport.util.exception.ConflictException;
 import org.folio.kafka.KafkaConfig;
 import org.folio.kafka.KafkaHeaderUtils;
 import org.folio.kafka.ProcessRecordErrorHandler;
@@ -37,10 +38,14 @@ public class StoredRecordChunksErrorHandler implements ProcessRecordErrorHandler
   public static final String RECORD_ID_HEADER = "recordId";
   private static final String CHUNK_NUMBER = "chunkNumber";
 
+  private final Vertx vertx;
+  private final KafkaConfig kafkaConfig;
+
   @Autowired
-  private Vertx vertx;
-  @Autowired
-  private KafkaConfig kafkaConfig;
+  public StoredRecordChunksErrorHandler(Vertx vertx, KafkaConfig kafkaConfig) {
+    this.vertx = vertx;
+    this.kafkaConfig = kafkaConfig;
+  }
 
   @Override
   public void handle(Throwable throwable, KafkaConsumerRecord<String, String> kafkaConsumerRecord) {
@@ -55,6 +60,9 @@ public class StoredRecordChunksErrorHandler implements ProcessRecordErrorHandler
       for (Record failedRecord: failedRecords) {
         sendDiErrorForRecord(jobExecutionId, failedRecord, okapiParams, failedRecord.getErrorRecord().getDescription());
       }
+
+    } else if (throwable instanceof ConflictException) {
+        LOGGER.warn(throwable.getMessage());
 
     } else {
       // process for all other cases that will include all records